requirements.txt
streamlit==1.28.0
plotly==5.17.0
pandas==2.0.3
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
chromadb==0.4.15
sentence-transformers==2.2.2
groq==0.4.1
streamlit-folium==0.15.0
python-dotenv==1.0.0